DNM: Add edpm_lvmfilter role
This is a port of the corresponding role in tripleo-ansible.
However, after porting much of the code to edpm-ansible, it became
apparent the functionality is not necessary in greenfiled deployments
where EDPM nodes are running EL9. This is due to enhancements in
the lvm2 package to add support for lvmdevices [1], which was
introduced in rhel-8.6 [2].

[1] https://man7.org/linux/man-pages/man8/lvmdevices.8.html
[2] https://bugzilla.redhat.com/show_bug.cgi?id=2012830

In short, it's no longer necessary to configure an LVM filter. In
fact, the global_filter section of lvm.conf is explicitly ignored
when lvmdevices are used.